Startup and general option handling
The idea is to exchange TTYPE options, and if it's a terminal then
we also exchange a bunch of other options.

We do that in parallel so that ideally the delay this causes is a
half roundtrip.

XXX this may trigger charset negotiation which we may or may not yet wait for.
We should.